refactor(dccpeep): extract label and branch pass family
Move pass_labels, branch-over-jump, jump threading, conditional shortcutting, and jump-to-plain-ret folding with their private rewrite helper into peep_pass_control_flow.c. Keep is_label_referenced exported because dead-label analysis also consumes it.

Validation: strict build clean; six fixtures pass; all 316 raw inputs byte-identical in three modes.
